--- back/src/main/java/capstone/facefriend/auth/controller/dto/TokenResponse.java
@@ -1,8 +1,0 @@
-package capstone.facefriend.auth.controller.dto;
-
-public record TokenResponse(
-        String accessToken,
-        String refreshToken,
-        Long memberId
-) {
-}
