+++ back/src/main/java/capstone/facefriend/auth/controller/interceptor/TokenReissueInterceptor.java
@@ -2,7 +2,7 @@ package capstone.facefriend.auth.controller.interceptor;
 
 import capstone.facefriend.auth.controller.support.AuthenticationContext;
 import capstone.facefriend.auth.controller.support.AuthenticationExtractor;
-import capstone.facefriend.auth.domain.TokenProvider;
+import capstone.facefriend.auth.domain.token.TokenProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
