+++ back/src/main/java/capstone/facefriend/auth/domain/oauth/GoogleMember.java
@@ -1,4 +1,4 @@
-package capstone.facefriend.auth.domain;
+package capstone.facefriend.auth.domain.oauth;
 
 import java.util.Map;
