+++ back/src/main/java/capstone/facefriend/auth/domain/token/AccessToken.java
@@ -1,0 +1,7 @@
+package capstone.facefriend.auth.domain.token;
+
+public record AccessToken(String value) {
+    public static AccessToken from(String value) {
+        return new AccessToken(value);
+    }
+}
