+++ back/src/main/java/capstone/facefriend/auth/domain/token/RefreshToken.java
@@ -1,0 +1,7 @@
+package capstone.facefriend.auth.domain.token;
+
+public record RefreshToken(String value) {
+    public static RefreshToken from(String value) {
+        return new RefreshToken(value);
+    }
+}
