+++ back/src/main/java/capstone/facefriend/auth/dto/LoginUriResponse.java
@@ -1,4 +1,4 @@
-package capstone.facefriend.auth.controller.dto;
+package capstone.facefriend.auth.dto;
 
 public record LoginUriResponse(
         String loginUri
