+++ back/src/main/java/capstone/facefriend/auth/service/OAuthRequester.java
@@ -1,8 +1,8 @@
 package capstone.facefriend.auth.service;
 
-import capstone.facefriend.auth.domain.OAuthMember;
-import capstone.facefriend.auth.domain.Provider;
-import capstone.facefriend.auth.service.dto.OAuthLoginRequest;
+import capstone.facefriend.auth.domain.oauth.OAuthMember;
+import capstone.facefriend.auth.domain.oauth.Provider;
+import capstone.facefriend.auth.dto.OAuthLoginRequest;
 
 public interface OAuthRequester {
